+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/ImpTest.hs
@@ -46,5 +46,3 @@ instance
   ShelleyEraImp (AlonzoEra c)
   where
   emptyImpNES = emptyAlonzoImpNES
-
-  impWitsVKeyNeeded = shelleyImpWitsVKeyNeeded